functions/requirements.txt
requests==2.31.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-scheduler==2.12.0
python-ta-lib==0.4.24
google-generativeai==0.3.1
pytz==2023.3
websocket-client==1.6.1
